requirements.txt
fastapi
uvicorn
passlib
bcrypt
pymongo[srv]
pytest
pyYAML
docker-compose
pydantic
coverage
pylint
flake8